zthumbor: Start using https_loader instead of http_loader.
https_loader makes use of https by default but fallbacks to http
when urls does not have https support. Thumbor's in built loaders
doesn't support this kind of fallback so we do a little bit of
due diligence on the results the in built loader was going to
throw at the callback and again request for image with a http_url
in case the result said that request for external type image failed.

thumbnail: Add thumbnail endpoint for passing requests to thumbor.

thumbnail: Start using thumbor for thumbnailing user uploads.

thumbnails: Make request signing process more robust.

We start to use the inbuilt request signing process of thumbor.
For this we introduce a new dependancy (the one used by thumbor for
generating sign hashes) and use it to generate hashes. We also
include the source_type in signing process by making it a part of
file path. We later on separate out source_type from file path.

thumbnails: Fix bug with thumbnails and S3 storage.

thumbnails: Make thumbnailing process more robust to DoS attacks.

TODOS:
* Change THUBMOR_HOST
* Debug the nginx issue from zerver/lib/thumbnail.py.
* Provide better error pages for unauthorized access.
* Add a test for accessing /thumbnail/ using the API.
* S3 is broken; we should be using get_secret to get the keys in thumbor.conf.
* the convert https to http function loses query parameters