thumbor: Add ldap and django as a dependancy.
We add this dependancy to thumbor for no use other than
making an import possible in one of the upcoming commits. Basically
we wanted to import LOCAL_UPLOADS_DIR from zproject.prod_settings or
zproject.dev_settings and these settings have use of LDAP library
functions. By doing this addition we are just adding some
packages to thumbor venv which we are not really gonna use so
it might feel a bit odd but it is ohk until it helps us reduce the
server restart time by 1s.